# src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: Aeza Server API
  version: 1.0.0
  description: REST, WebSocket and SSE API for frontend and agents.
servers:
  - url: http://158.160.205.79:8087
    description: Current deployed server
  - url: /
    description: Relative current server
paths:
...
  /api/v1/agents:
    get:
      summary: List agents
      description: Returns current list of agents for frontend. Can be filtered by groupId.
      parameters:
        - in: query
          name: groupId
          required: false
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Agent list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AgentStatus"
              example:
                - agentId: "agent-dev-01"
                  name: "Main dev agent"
                  host: "agent-vm-01"
                  hostIp: "158.160.205.79"
                  status: "online"
                  lastSeen: "2026-03-21T11:24:32.636001841Z"
                  bufferedCount: 0
                  groupId: 1
                  groupName: "production"
    post:
      summary: Create agent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentCreateRequest"
            example:
              agentId: "agent-prod-02"
              name: "Prod node 02"
              host: "prod-node-02"
              hostIp: "10.0.0.42"
              status: "offline"
              bufferedCount: 0
              groupId: 1
      responses:
        "200":
          description: Created agent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentStatus"
  /api/v1/agents/{id}:
    get:
      summary: Get agent details
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Agent details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentStatus"
              example:
                agentId: "agent-dev-01"
                name: "Main dev agent"
                host: "agent-vm-01"
                hostIp: "158.160.205.79"
                status: "online"
                lastSeen: "2026-03-21T11:24:32.636001841Z"
                bufferedCount: 0
                groupId: 1
                groupName: "production"
        "404":
          description: Agent not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
              example:
                timestamp: "2026-03-21T12:00:00Z"
                status: 404
                error: "Not Found"
                message: "agent not found"
                path: "/api/v1/agents/agent-404"
    put:
      summary: Update agent
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentUpdateRequest"
            example:
              name: "Prod node 02 renamed"
              status: "online"
              groupId: 2
      responses:
        "200":
          description: Updated agent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentStatus"
  /api/v1/agent-groups:
    get:
      summary: List agent groups
      responses:
        "200":
          description: Agent groups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AgentGroup"
    post:
      summary: Create agent group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentGroupCreateRequest"
            example:
              name: "production"
              description: "Main production servers"
              agentIds:
                - "agent-prod-01"
                - "agent-prod-02"
      responses:
        "200":
          description: Created agent group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentGroup"
  /api/v1/agent-groups/{id}:
    get:
      summary: Get agent group details
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Agent group details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentGroup"
    put:
      summary: Update agent group
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentGroupUpdateRequest"
            example:
              name: "production-core"
              description: "Core production nodes"
              agentIds:
                - "agent-prod-01"
                - "agent-prod-02"
                - "agent-prod-03"
      responses:
        "200":
          description: Updated agent group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentGroup"
    delete:
      summary: Delete agent group
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: Group deleted
...
components:
  schemas:
...
    AgentStatus:
      type: object
      properties:
        agentId:
          type: string
        name:
          type: string
        host:
          type: string
        hostIp:
          type: string
          nullable: true
        status:
          type: string
        lastSeen:
          type: string
          format: date-time
        bufferedCount:
          type: integer
        groupId:
          type: integer
          format: int64
          nullable: true
        groupName:
          type: string
          nullable: true
    AgentCreateRequest:
      type: object
      required:
        - agentId
        - host
      properties:
        agentId:
          type: string
        name:
          type: string
        host:
          type: string
        hostIp:
          type: string
          nullable: true
        status:
          type: string
        bufferedCount:
          type: integer
          minimum: 0
        groupId:
          type: integer
          format: int64
          nullable: true
    AgentUpdateRequest:
      type: object
      properties:
        name:
          type: string
        host:
          type: string
        hostIp:
          type: string
          nullable: true
        status:
          type: string
        bufferedCount:
          type: integer
          minimum: 0
        groupId:
          type: integer
          format: int64
          nullable: true
        clearGroup:
          type: boolean
    AgentGroupCreateRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        agentIds:
          type: array
          items:
            type: string
    AgentGroupUpdateRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        agentIds:
          type: array
          items:
            type: string
    AgentGroup:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        agentCount:
          type: integer
          format: int64
        agents:
          type: array
          items:
            $ref: "#/components/schemas/AgentStatus"
...
    ApiErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string